graph-substitution.ts
