+++ aa.js
@@ -1,0 +1,3 @@
+var a;
+var b = (a = 3) ? true : false;
+console.log(b);
